Get the public disease list

'get_public_disease_list' reads the public list from a given CSV file or uses the default diseases if the file doesn't exist.

Usage

get_public_disease_list(filepath, default_diseases)

Arguments

filepath

Filepath. Public disease list CSV file.

default_diseases

String vector. List of default diseases to use if the above file doesn't exist.

Value

A dataframe containing the diseases to include in the public report and the name to use for each disease in the public report.

Details

The provided public disease list file must contain two columns that map the EpiTrax disease name to a public-facing name for the public report.

Examples

if (FALSE) { # \dontrun{
  list_file <- "path/to/file"
  default_list <- c("Measles", "Chickenpox")

  disease_list <- get_public_disease_list(list_file, default_list)
} # }